requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
google-generativeai==0.8.3
google-cloud-aiplatform==1.74.0
chromadb==0.5.23
PyMuPDF==1.25.1
sentence-transformers==3.3.1
python-dotenv==1.0.1
pydantic==2.10.3
requests==2.32.3
redis==5.2.1
rank-bm25==0.2.2
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.0.2
pytest==8.3.4
httpx==0.28.1
locust==2.32.6
